Refactor opengl examples to reduce code duplication and fix warnings
